start-qemu.sh: allow to specify number of CPUs via cmdline
The number of CPUs is presently hardcoded as 1.
To change the value, user needs to edit the shell script.

This patch adds support to pass the desired number of CPUs via a new
command line option "-c <number>". The default value remains "1" as before.

Number of CPUs must be positive, there are no upper limit checks.
The assumption here is that the user knows what he/she is doing.

Example: Create VM with 16 CPUs:
 ./start-qemu.sh -i td-guest-rhel8.5.qcow2 -k vmlinuz-rhel8.5 -c 16

The number of CPUs can be verified in VM via the command 'lscpu'.

This patch contains additional changes:
- It removes the variable SOCKET, there is no useful user scenario when one
  would want to use non-default number of sockets "1".
- The variable CORES was renamed to CPUS.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>