build-repo.sh: update with SPR kernel
Since the SPR kernel spec builds both the host and guest packages the
build-repo script needed to be modified to take this into account. A
new function was added to move packages from one repo to another if
necessary. The final createrepo_c step was moved to its own finalize
function as it didn't make sense for this to happen before moving
packages to their proper locations.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>