build: remove rhel-8 pkg-builder
RHEL is a commercial distro and only provides the base packages for
their public facing container. We currently don't have a means to
import the subscription service into the container, which would be
needed to install the package build requirements.

Since the pkg-builder is not strictly needed to build the packages,
drop it for now.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>