pycloudstack/virtxml.py: memory.setter: handle float instance
Running the regression suit of tests:
  sudo ./run.sh -s regression

typically results in several tests failing with an error such as:
  text = 16777216.0
  ...
  TypeError: argument of type 'float' is not iterable

This patch detects attempts setting memory field in XML using 'float'.
The 'float' value is first converted to 'int' using 'round' to eliminate
decimal values and then the 'int' value is converted to a string.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>