Bug fix: fails to correctly detect TDX guest
Updated line 7 of the script to be able to correctly detect TDX guests. In some platforms where the string "tdx_guest" appears multiple times as the output of "grep -o tdx_guest /proc/cpuinfo", earlier comparison that checked equality failed. 

This fix will solve that bug and the script will be able to correct detect a TDX guest in those platforms as well as others.